Move apiRouter into routers
